Started work on graphical effect for abilities and weapon
Also State in weapon is no running on server and client, though it is
not really synced, and server still have full authority, client just
"simulate" what is going on, but doesn't tell server the results of
simulation and server results are authorative at any point. Which might
cause desync, where there is big lag or client is trying to cheat.