Moving all attribute handling to AttributeBaseComp
All events, replication, and changing of attribute will be handled
trough AttributeBaseComponent. If needed this component can be attached
to controller to get attribute changes event higher in controll
hierarchy.

It's going to help with maintanability and extensibility. Now that
everything is sitting in single component, it can be attached to any
actor.