Added custom asset type
Custom asset type, which will hold information about items.
Next step refactoring, inventory and equipment system to take advantage
of this asset,

Editor is available, and should work for most part.
Keys to assets are not really readable by humans (they are GUIDs), and
should be referenced directly.

In future there should be another extension to editor, which will show
list of item to select from selected asset.

The asset format is not final, and can be easily extended or changed.
All that is needed, is to chage structs inside ARItemsData.h (and places
where those structs are referenced, though i tried to keep it to
minimum) and override operator<< for FArchive in ActionRPGGame.h to
accomodate for struct changes